init-scripts: use start-stop-daemon
Before, we started all daemon process on our own.
It is more convenient and error-proof if we rely on the
existing util for this task (start-stop-daemon).
This path introduces usage of start-stop-daemon for
all our daemons.

This also removes the automated redirection of the stdout and
stderr from all started daemons. Effectively, this removes all
logging from our processes since the stdout and stderr of daemons
is redirected to /dev/null. Therefore, we will have to
introduce proper logging with syslogd in the upcoming patches.
logging to syslogd