init-scripts: ensure processes are not started twice
Usually, start-stop-daemon makes sure that the same process
is not started twice. However, with the bash scripts that we start,
the mechanism does not function. We can start the same process
with the same start-stop-daemon command twice.
In order to prevent this behavior, we add an additional command
that checks for a living process with the given PID before
starting a new instance.