Make sure that gpio scripts kill child processes
Before this commit, killing the bash script did not kill the
child process gpio-notify. In order to also kill the child process,
in this patch, we set up a trap, send the child process to the
background and wait for it (wait is interruptible by sinals).
In addition, since we want to capture the output of the
gpio-notify process (that runs in the background now), we
send the output to a pipe and read from it in the main process.
This gives us the output of the process.